feat: добавил рисование линии

+++ Lab4/Lab4_1/Lab4_1/Core/Abstraction/ICanvas.cs
@@ -7,10 +7,10 @@ public interface ICanvas
     void DrawLine(Point start, Point finish, uint lineColor);
     void DrawCircle(Point center, double radius, uint lineColor);
     void FillCircle(Point center, double radius, uint fillColor);
-    void DrawTriangle(Point firstPoint, Point secondPoint, Point thirdPoint, double lineColor);
-    void FillTriangle(Point firstPoint, Point secondPoint, Point thirdPoint, double fillColor);
-    void DrawRectangle(Point firstPoint, double width, double height, double lineColor);
-    void FillRectangle(Point firstPoint, double width, double height, double fillColor);
+    void DrawTriangle(Point firstPoint, Point secondPoint, Point thirdPoint, uint lineColor);
+    void FillTriangle(Point firstPoint, Point secondPoint, Point thirdPoint, uint fillColor);
+    void DrawRectangle(Point firstPoint, double width, double height, uint lineColor);
+    void FillRectangle(Point firstPoint, double width, double height, uint fillColor);
 
     void Clear();
     // void
